fix(document): serialize tag value in the correct priority
