+++ docarray/array/document.py
@@ -32,6 +32,21 @@ if TYPE_CHECKING:
 
 
 class DocumentArray(AllMixins, MutableSequence[Document]):
+    def _get_doc_by_offset(self, offset: int) -> 'Document':
+        return self._data[offset]
+
+    def _get_doc_by_id(self, _id: str) -> 'Document':
+        return self._data[self._id2offset[_id]]
+
+    def _get_docs_by_slice(self, _slice: slice) -> Iterable['Document']:
+        return self._data[_slice]
+
+    def _get_docs_by_offsets(self, offsets: Sequence[int]) -> Iterable['Document']:
+        return (self._data[t] for t in offsets)
+
+    def _get_docs_by_ids(self, ids: Sequence[str]) -> Iterable['Document']:
+        return (self._data[self._id2offset[t]] for t in ids)
+
     def __init__(
         self, docs: Optional['DocumentArraySourceType'] = None, copy: bool = False
     ):
@@ -107,72 +122,6 @@ class DocumentArray(AllMixins, MutableSequence[Document]):
             return x.id in self._id2offset
         else:
             return False
-
-    @overload
-    def __getitem__(self, index: 'DocumentArraySingletonIndexType') -> 'Document':
-        ...
-
-    @overload
-    def __getitem__(self, index: 'DocumentArrayMultipleIndexType') -> 'DocumentArray':
-        ...
-
-    @overload
-    def __getitem__(self, index: 'DocumentArraySingleAttributeType') -> List[Any]:
-        ...
-
-    @overload
-    def __getitem__(
-        self, index: 'DocumentArrayMultipleAttributeType'
-    ) -> List[List[Any]]:
-        ...
-
-    def __getitem__(
-        self, index: 'DocumentArrayIndexType'
-    ) -> Union['Document', 'DocumentArray']:
-        if isinstance(index, (int, np.generic)) and not isinstance(index, bool):
-            return self._data[int(index)]
-        elif isinstance(index, str):
-            if index.startswith('@'):
-                return self.traverse_flat(index[1:])
-            else:
-                return self._data[self._id2offset[index]]
-        elif isinstance(index, slice):
-            return DocumentArray(self._data[index])
-        elif index is Ellipsis:
-            return self.flatten()
-        elif isinstance(index, Sequence):
-            if (
-                isinstance(index, tuple)
-                and len(index) == 2
-                and isinstance(index[0], (slice, Sequence))
-            ):
-                if isinstance(index[0], str) and isinstance(index[1], str):
-                    # ambiguity only comes from the second string
-                    if index[1] in self:
-                        return DocumentArray([self[index[0]], self[index[1]]])
-                    else:
-                        return getattr(self[index[0]], index[1])
-                elif isinstance(index[0], (slice, Sequence)):
-                    _docs = self[index[0]]
-                    _attrs = index[1]
-                    if isinstance(_attrs, str):
-                        _attrs = (index[1],)
-                    return _docs._get_attributes(*_attrs)
-            elif isinstance(index[0], bool):
-                return DocumentArray(itertools.compress(self._data, index))
-            elif isinstance(index[0], int):
-                return DocumentArray(self[t] for t in index)
-            elif isinstance(index[0], str):
-                return DocumentArray(self[t] for t in index)
-        elif isinstance(index, np.ndarray):
-            index = index.squeeze()
-            if index.ndim == 1:
-                return self[index.tolist()]
-            else:
-                raise IndexError(
-                    f'When using np.ndarray as index, its `ndim` must =1. However, receiving ndim={index.ndim}'
-                )
-        raise IndexError(f'Unsupported index type {typename(index)}: {index}')
 
     @overload
     def __setitem__(
